selectErrorBarsSettings only returns the ErrorBars that are relevant for the selected axis
