+++ src/cartesian/Area.tsx
@@ -612,6 +612,7 @@ export class Area extends PureComponent<Props, State> {
           <SetCartesianGraphicalItem
             data={this.props.data}
             xAxisId={this.props.xAxisId}
+            yAxisId={this.props.yAxisId}
             dataKey={this.props.dataKey}
             errorBars={noErrorBars}
           />
@@ -633,7 +634,12 @@ export class Area extends PureComponent<Props, State> {
     const dotSize = r * 2 + strokeWidth;
 
     return (
-      <CartesianGraphicalItemContext data={this.props.data} xAxisId={this.props.xAxisId} dataKey={this.props.dataKey}>
+      <CartesianGraphicalItemContext
+        data={this.props.data}
+        xAxisId={this.props.xAxisId}
+        dataKey={this.props.dataKey}
+        yAxisId={this.props.yAxisId}
+      >
         <Layer className={layerClass}>
           <SetAreaLegend {...this.props} />
           <SetTooltipEntrySettings fn={getTooltipEntrySettings} args={this.props} />
